feat(table)!: make it responsive and look better (see desc)
It was made responsive by hiding table header columns (th tags) and showing details only (for smaller screens). It was implemented using Tailwind CSS classes, regular CSS and useMediaQuery hook from @mantine/hooks. In this commit there is also a lot of minor changes to the layout to make it visually appealing. The biggest struggle in the process was to make it work with previously implemented infinite scroll with query pagination.